+++ modules/core/02-client/types/keys.go
@@ -32,11 +32,6 @@ const (
 	// AllowAllClients is the value that if set in AllowedClients param
 	// would allow any wired up light client modules to be allowed
 	AllowAllClients = "*"
-
-	// CounterpartyKey is the key used to store counterparty in the client store.
-	// the counterparty key is imported from types instead of host because
-	// the counterparty key is not a part of the ics-24 host specification
-	CounterpartyKey = "counterparty"
 )
 
 // FormatClientIdentifier returns the client identifier with the sequence appended.
