+++ modules/core/04-channel/v2/module.go
@@ -7,17 +7,17 @@ import (
 	"github.com/cosmos/ibc-go/v9/modules/core/04-channel/v2/types"
 )
 
-// Name returns the IBC client name
+// Name returns the IBC channel/v2 name
 func Name() string {
 	return types.SubModuleName
 }
 
-// GetQueryCmd returns no root query command for the IBC client
+// GetQueryCmd returns the root query command for IBC channels v2.
 func GetQueryCmd() *cobra.Command {
 	return cli.GetQueryCmd()
 }
 
-// GetTxCmd returns the root tx command for 02-client.
+// GetTxCmd returns the root tx command for IBC channels v2.
 func GetTxCmd() *cobra.Command {
 	return cli.NewTxCmd()
 }
