+++ modules/core/keeper/msg_server.go
@@ -19,10 +19,9 @@ import (
 )
 
 var (
-	_ clienttypes.MsgServer        = (*Keeper)(nil)
-	_ connectiontypes.MsgServer    = (*Keeper)(nil)
-	_ channeltypes.MsgServer       = (*Keeper)(nil)
-	_ channeltypes.PacketMsgServer = (*Keeper)(nil)
+	_ clienttypes.MsgServer     = (*Keeper)(nil)
+	_ connectiontypes.MsgServer = (*Keeper)(nil)
+	_ channeltypes.MsgServer    = (*Keeper)(nil)
 )
 
 // CreateClient defines a rpc handler method for MsgCreateClient.
